1, Fix new list problem.

+++ app/assets/javascripts/administration/v1/new/index.js
@@ -5,26 +5,31 @@
 
 $(function(){
 	$('.js-delete-one-new').on('click',function(){
-		$.ajax({
-			url: "/administration-api/v1/new/"+$(this).attr('data-new-id'),
-			type: "delete",
-			data: {
-
-			},
-			dataType: "json"
-		})
-			.done(function( data ) {
-				if(data['success']){
-					window.location.href='/administration/v1/new';
-				}else{
-					alert(data['message']);
-				}
-			})
-			.fail(function( xhr, status, errorThrown ) {
-				alert('异常');
-				return false;
-			})
-			.always(function( xhr, status ) {
-			});
+   var newId = $(this).attr('data-new-id');
+		AMUI.dialog.confirm({
+			title: '警告',
+			content: '确认删除吗？',
+			onConfirm: function() {
+				$.ajax({
+					url: "/administration-api/v1/new/"+newId,
+					type: "delete",
+					data: {
+					},
+					dataType: "json"
+				})
+					.done(function( data ) {
+						if(data['success']){
+							window.location.href='/administration/v1/new';
+						}else{
+							alert(data['message']);
+						}
+					})
+					.fail(function( xhr, status, errorThrown ) {
+						console.log(xhr['responseJSON']);
+						return false;
+					})
+					.always(function( xhr, status ) {
+					});
+		 }, onCancel: function() {  } });
 	});
 })
